+++ example/test/examples/motion_example_test.dart
@@ -18,4 +18,20 @@ void main() {
       expect(find.text('Motion completed'), findsOneWidget);
     },
   );
+
+  testWidgets(
+    'when the replay action is tapped, it should start the Motion again',
+    (tester) async {
+      await tester.pumpWidget(
+        const MaterialApp(home: Scaffold(body: MotionExample())),
+      );
+      await tester.pump(const Duration(seconds: 1));
+      await tester.tap(find.text('Play motion again'));
+      await tester.pump();
+      await tester.pump(const Duration(milliseconds: 100));
+      await tester.pump();
+
+      expect(find.text('Motion started'), findsOneWidget);
+    },
+  );
 }